+++ src/main/java/com/mraof/minestuck/block/machine/AnthvilBlock.java
@@ -1,0 +1,58 @@
+package com.mraof.minestuck.block.machine;
+
+import com.mraof.minestuck.blockentity.MSBlockEntityTypes;
+import com.mraof.minestuck.blockentity.machine.AnthvilBlockEntity;
+import net.minecraft.core.BlockPos;
+import net.minecraft.server.level.ServerPlayer;
+import net.minecraft.world.InteractionHand;
+import net.minecraft.world.InteractionResult;
+import net.minecraft.world.MenuProvider;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.level.Level;
+import net.minecraft.world.level.block.EntityBlock;
+import net.minecraft.world.level.block.entity.BlockEntity;
+import net.minecraft.world.level.block.entity.BlockEntityTicker;
+import net.minecraft.world.level.block.entity.BlockEntityType;
+import net.minecraft.world.level.block.state.BlockState;
+import net.minecraft.world.phys.BlockHitResult;
+import net.minecraftforge.network.NetworkHooks;
+
+import javax.annotation.Nullable;
+
+public class AnthvilBlock extends MachineProcessBlock implements EntityBlock
+{
+	public AnthvilBlock(Properties properties)
+	{
+		super(properties);
+	}
+	
+	@Override
+	@SuppressWarnings("deprecation")
+	public InteractionResult use(BlockState state, Level level, BlockPos pos, Player player, InteractionHand handIn, BlockHitResult hit)
+	{
+		if(player instanceof ServerPlayer serverPlayer)
+		{
+			BlockEntity blockEntity = level.getBlockEntity(pos);
+			if(blockEntity instanceof AnthvilBlockEntity)
+			{
+				MenuProvider menuProvider = (MenuProvider) blockEntity;
+				NetworkHooks.openScreen(serverPlayer, menuProvider, pos);
+			}
+		}
+		return InteractionResult.sidedSuccess(level.isClientSide);
+	}
+	
+	@Nullable
+	@Override
+	public BlockEntity newBlockEntity(BlockPos pos, BlockState state)
+	{
+		return new AnthvilBlockEntity(pos, state);
+	}
+	
+	@Nullable
+	@Override
+	public <E extends BlockEntity> BlockEntityTicker<E> getTicker(Level level, BlockState state, BlockEntityType<E> placedType)
+	{
+		return createMachineTicker(level, placedType, MSBlockEntityTypes.ANTHVIL.get());
+	}
+}