+++ src/main/java/com/mraof/minestuck/data/MinestuckEnUsLanguageProvider.java
@@ -528,6 +528,8 @@ public class MinestuckEnUsLanguageProvider extends MinestuckLanguageProvider
 		addBlock(MSBlocks.URANIUM_COOKER, "Cookalyzer");
 		addBlock(MSBlocks.GRIST_COLLECTOR, "Grist Collector");
 		addBlockTooltip(MSBlocks.GRIST_COLLECTOR, "Collects grist in a one block radius and releases it when right clicked.");
+		addBlock(MSBlocks.ANTHVIL, "Anthvil");
+		addBlockTooltip(MSBlocks.ANTHVIL, "Recover lost durability on weapons that have a grist cost.");
 		addBlock(MSBlocks.SKAIANET_DENIER, "Skaianet Denier");
 		addBlockTooltip(MSBlocks.SKAIANET_DENIER, "Constantly sends out signals preventing Entry from occurring while it is inside the radius.");
 		addBlock(MSBlocks.CRUXITE_DOWEL, "Cruxite Dowel");
@@ -1744,6 +1746,7 @@ public class MinestuckEnUsLanguageProvider extends MinestuckLanguageProvider
 		add(UraniumCookerBlockEntity.TITLE, "Cookalyzer");
 		add(SendificatorBlockEntity.TITLE, "Sendificator");
 		add(GristWidgetBlockEntity.TITLE, "GristWidget 12000");
+		add(AnthvilBlockEntity.TITLE, "Anthvil");
 		add(RedstoneClockBlockEntity.TIME_CHANGE, "The clock now pulses every %s seconds");
 		add(GuiUtil.NOT_ALCHEMIZABLE, "Not Alchemizable");
 		add(GuiUtil.FREE, "Free!");