+++ CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased]
 
+### Added
+
+- Anthvil block
+
+### Contributors for this release
+
+- rose_bushes_, Dweblenod
+
 ## [1.19.2-1.10.2.0] - 2023-04-20
 
 ### Added